backend/requirements.txt
# Updated requirements compatible with Python 3.14
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
httpx>=0.27.0

# Auth & utilities
PyJWT>=2.8.0
geopy>=2.4.0
numpy>=1.26.0

# Google AI
google-generativeai>=0.5.0

# Supabase
supabase>=2.4.1

# Build tools
setuptools
wheel
